renamed bundle as wished in pr
