+++ bundles/customer-anonymizer-company-user-connector/phpstan.neon
