+++ bundles/customer-anonymizer-company-user-connector/src/FondOfImpala/Zed/CustomerAnonymizerCompanyUserConnector/Business/CustomerAnonymizerCompanyUserConnectorFacade.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace FondOfImpala\Zed\CustomerAnonymizerCompanyUserConnector\Business;
+
+use Generated\Shared\Transfer\CustomerTransfer;
+use Spryker\Zed\Kernel\Business\AbstractFacade;
+
+/**
+ * @method \FondOfImpala\Zed\CustomerAnonymizerCompanyUserConnector\Business\CustomerAnonymizerCompanyUserConnectorBusinessFactory getFactory()
+ */
+class CustomerAnonymizerCompanyUserConnectorFacade extends AbstractFacade implements CustomerAnonymizerCompanyUserConnectorFacadeInterface
+{
+    /**
+     * @param \Generated\Shared\Transfer\CustomerTransfer $customerTransfer
+     *
+     * @return void
+     */
+    public function deleteCompanyUsersForCustomer(CustomerTransfer $customerTransfer): void
+    {
+        $this->getFactory()->createCompanyUserDeleter()->deleteByCustomer($customerTransfer);
+    }
+}
