+++ bundles/customer-anonymizer-company-user-connector/src/FondOfImpala/Zed/CustomerAnonymizerCompanyUserConnector/Business/CustomerAnonymizerCompanyUserConnectorFacadeInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace FondOfImpala\Zed\CustomerAnonymizerCompanyUserConnector\Business;
+
+use Generated\Shared\Transfer\CustomerTransfer;
+
+interface CustomerAnonymizerCompanyUserConnectorFacadeInterface
+{
+    /**
+     * @param \Generated\Shared\Transfer\CustomerTransfer $customerTransfer
+     *
+     * @return void
+     */
+    public function deleteCompanyUsersForCustomer(CustomerTransfer $customerTransfer): void;
+}
