+++ bundles/customer-anonymizer-company-user-connector/src/FondOfImpala/Zed/CustomerAnonymizerCompanyUserConnector/Business/Model/CompanyUserDeleterInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+declare (strict_types=1);
+
+namespace FondOfImpala\Zed\CustomerAnonymizerCompanyUserConnector\Business\Model;
+
+use Generated\Shared\Transfer\CustomerTransfer;
+
+interface CompanyUserDeleterInterface
+{
+    /**
+     * @param \Generated\Shared\Transfer\CustomerTransfer $customerTransfer
+     *
+     * @return void
+     */
+    public function deleteByCustomer(CustomerTransfer $customerTransfer): void;
+}
