+++ bundles/customer-anonymizer-company-user-connector/src/FondOfImpala/Zed/CustomerAnonymizerCompanyUserConnector/CustomerAnonymizerCompanyUserConnectorDependencyProvider.php
@@ -1,0 +1,80 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace FondOfImpala\Zed\CustomerAnonymizerCompanyUserConnector;
+
+use FondOfImpala\Zed\CustomerAnonymizerCompanyUserConnector\Dependency\Facade\CustomerAnonymizerCompanyUserConnectorToCompanyUserFacadeBridge;
+use Orm\Zed\CompanyUser\Persistence\SpyCompanyUserQuery;
+use Spryker\Zed\Kernel\AbstractBundleDependencyProvider;
+use Spryker\Zed\Kernel\Container;
+
+/**
+ * @codeCoverageIgnore
+ */
+class CustomerAnonymizerCompanyUserConnectorDependencyProvider extends AbstractBundleDependencyProvider
+{
+    /**
+     * @var string
+     */
+    public const FACADE_COMPANY_USER = 'FACADE_COMPANY_USER';
+
+    /**
+     * @var string
+     */
+    public const QUERY_COMPANY_USER = 'QUERY_COMPANY_USER';
+
+    /**
+     * @param \Spryker\Zed\Kernel\Container $container
+     *
+     * @return \Spryker\Zed\Kernel\Container
+     */
+    public function provideBusinessLayerDependencies(Container $container): Container
+    {
+        $container = parent::provideBusinessLayerDependencies($container);
+
+        return $this->addCompanyUserFacade($container);
+    }
+
+    /**
+     * @param \Spryker\Zed\Kernel\Container $container
+     *
+     * @return \Spryker\Zed\Kernel\Container
+     */
+    public function providePersistenceLayerDependencies(Container $container): Container
+    {
+        $container = parent::providePersistenceLayerDependencies($container);
+
+        return $this->addCompanyUserQuery($container);
+    }
+
+    /**
+     * @param \Spryker\Zed\Kernel\Container $container
+     *
+     * @return \Spryker\Zed\Kernel\Container
+     */
+    protected function addCompanyUserFacade(Container $container): Container
+    {
+        $container[static::FACADE_COMPANY_USER] = static fn (
+            Container $container
+        ): CustomerAnonymizerCompanyUserConnectorToCompanyUserFacadeBridge => new CustomerAnonymizerCompanyUserConnectorToCompanyUserFacadeBridge(
+            $container->getLocator()->companyUser()->facade(),
+        );
+
+        return $container;
+    }
+
+    /**
+     * @param \Spryker\Zed\Kernel\Container $container
+     *
+     * @return \Spryker\Zed\Kernel\Container
+     */
+    protected function addCompanyUserQuery(Container $container): Container
+    {
+        $container[static::QUERY_COMPANY_USER] = static fn (
+            Container $container
+        ): SpyCompanyUserQuery => new SpyCompanyUserQuery();
+
+        return $container;
+    }
+}
