+++ bundles/customer-anonymizer-company-user-connector/src/FondOfImpala/Zed/CustomerAnonymizerCompanyUserConnector/Dependency/Facade/CustomerAnonymizerCompanyUserConnectorToCompanyUserFacadeInterface.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace FondOfImpala\Zed\CustomerAnonymizerCompanyUserConnector\Dependency\Facade;
+
+use Generated\Shared\Transfer\CompanyUserResponseTransfer;
+use Generated\Shared\Transfer\CompanyUserTransfer;
+
+interface CustomerAnonymizerCompanyUserConnectorToCompanyUserFacadeInterface
+{
+    /**
+     * Specification:
+     * - Executes CompanyUserPreDeletePluginInterface plugins before delete company user.
+     * - Deletes a company user.
+     *
+     * @api
+     *
+     * @param \Generated\Shared\Transfer\CompanyUserTransfer $companyUserTransfer
+     *
+     * @return \Generated\Shared\Transfer\CompanyUserResponseTransfer
+     */
+    public function deleteCompanyUser(CompanyUserTransfer $companyUserTransfer): CompanyUserResponseTransfer;
+}
