+++ bundles/customer-anonymizer-company-user-connector/src/FondOfImpala/Zed/CustomerAnonymizerCompanyUserConnector/Persistence/CustomerAnonymizerCompanyUserConnectorRepositoryInterface.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace FondOfImpala\Zed\CustomerAnonymizerCompanyUserConnector\Persistence;
+
+use Generated\Shared\Transfer\CompanyUserCollectionTransfer;
+
+interface CustomerAnonymizerCompanyUserConnectorRepositoryInterface
+{
+    /**
+     * @param int $fkCustomer
+     *
+     * @throws \Spryker\Zed\Kernel\Exception\Container\ContainerKeyNotFoundException
+     * @throws \Spryker\Zed\Propel\Business\Exception\AmbiguousComparisonException
+     *
+     * @return \Generated\Shared\Transfer\CompanyUserCollectionTransfer
+     */
+    public function findCompanyUsersByFkCustomer(int $fkCustomer): CompanyUserCollectionTransfer;
+}
