--- bundles/customer-company-user-connector/composer.json
@@ -1,40 +1,0 @@
-{
-  "name": "fond-of-impala/customer-company-user-connector",
-  "description": "Customer company user connector bundle",
-  "license": "MIT",
-  "authors": [
-    {
-      "name": "Julian Hyatt",
-      "email": "dev@jzit.de"
-    }
-  ],
-  "require": {
-    "php": ">=8.0",
-    "spryker/company-user": "^2.0.0",
-    "spryker/customer": "^7.0.0"
-  },
-  "require-dev": {
-    "fond-of-codeception/spryker": "*",
-    "spryker/code-sniffer": "*",
-    "spryker-sdk/phpstan-spryker": "*"
-  },
-  "minimum-stability": "dev",
-  "prefer-stable": true,
-  "autoload": {
-    "psr-4": {
-      "FondOfImpala\\": "src/FondOfImpala/"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "FondOfImpala\\": "tests/FondOfImpala/",
-      "Generated\\": "src/Generated/",
-      "Orm\\": "src/Orm/"
-    }
-  },
-  "config": {
-    "allow-plugins": {
-      "dealerdirect/phpcodesniffer-composer-installer": true
-    }
-  }
-}
