--- bundles/customer-company-user-connector/src/FondOfImpala/Zed/CustomerCompanyUserConnector/Business/CustomerCompanyUserConnectorFacadeInterface.php
@@ -1,17 +1,0 @@
-<?php
-
-declare(strict_types = 1);
-
-namespace FondOfImpala\Zed\CustomerCompanyUserConnector\Business;
-
-use Generated\Shared\Transfer\CustomerTransfer;
-
-interface CustomerCompanyUserConnectorFacadeInterface
-{
-    /**
-     * @param \Generated\Shared\Transfer\CustomerTransfer $customerTransfer
-     *
-     * @return void
-     */
-    public function deleteCompanyUserForCustomer(CustomerTransfer $customerTransfer): void;
-}
