--- bundles/customer-company-user-connector/src/FondOfImpala/Zed/CustomerCompanyUserConnector/Dependency/Facade/CustomerCompanyUserConnectorToCompanyUserFacadeInterface.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace FondOfImpala\Zed\CustomerCompanyUserConnector\Dependency\Facade;
-
-use Generated\Shared\Transfer\CompanyUserResponseTransfer;
-use Generated\Shared\Transfer\CompanyUserTransfer;
-
-interface CustomerCompanyUserConnectorToCompanyUserFacadeInterface
-{
-    /**
-     * Specification:
-     * - Executes CompanyUserPreDeletePluginInterface plugins before delete company user.
-     * - Deletes a company user.
-     *
-     * @api
-     *
-     * @param \Generated\Shared\Transfer\CompanyUserTransfer $companyUserTransfer
-     *
-     * @return \Generated\Shared\Transfer\CompanyUserResponseTransfer
-     */
-    public function deleteCompanyUser(CompanyUserTransfer $companyUserTransfer): CompanyUserResponseTransfer;
-}
