--- bundles/customer-company-user-connector/tests/FondOfImpala/Zed/CustomerCompanyUserConnector/CustomerCompanyUserConnectorDependencyProviderTest.php
@@ -1,110 +1,0 @@
-<?php
-
-namespace FondOfImpala\Zed\CustomerCompanyUserConnector;
-
-use Codeception\Test\Unit;
-use FondOfImpala\Zed\CustomerCompanyUserConnector\Dependency\Facade\CustomerCompanyUserConnectorToCompanyUserFacadeInterface;
-use Orm\Zed\CompanyUser\Persistence\SpyCompanyUserQuery;
-use PHPUnit\Framework\MockObject\MockObject;
-use Spryker\Shared\Kernel\BundleProxy;
-use Spryker\Zed\CompanyUser\Business\CompanyUserFacadeInterface;
-use Spryker\Zed\Kernel\Container;
-use Spryker\Zed\Kernel\Locator;
-
-class CustomerCompanyUserConnectorDependencyProviderTest extends Unit
-{
-    /**
-     * @var \Spryker\Zed\Kernel\Container|\PHPUnit\Framework\MockObject\MockObject
-     */
-    protected Container|MockObject $containerMock;
-
-    /**
-     * @var \PHPUnit\Framework\MockObject\MockObject|\Spryker\Zed\Kernel\Locator
-     */
-    protected MockObject|Locator $locatorMock;
-
-    /**
-     * @var \PHPUnit\Framework\MockObject\MockObject|\Spryker\Shared\Kernel\BundleProxy
-     */
-    protected MockObject|BundleProxy $bundleProxyMock;
-
-    /**
-     * @var \Spryker\Zed\CompanyUser\Business\CompanyUserFacadeInterface|\PHPUnit\Framework\MockObject\MockObject
-     */
-    protected CompanyUserFacadeInterface|MockObject $companyUserFacadeMock;
-
-    /**
-     * @var \FondOfImpala\Zed\CustomerCompanyUserConnector\CustomerCompanyUserConnectorDependencyProvider
-     */
-    protected CustomerCompanyUserConnectorDependencyProvider $dependencyProvider;
-
-    /**
-     * @return void
-     */
-    protected function _before(): void
-    {
-        $this->containerMock = $this->getMockBuilder(Container::class)
-            ->setMethodsExcept(['factory', 'set', 'offsetSet', 'get', 'offsetGet'])
-            ->getMock();
-
-        $this->locatorMock = $this->getMockBuilder(Locator::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $this->bundleProxyMock = $this->getMockBuilder(BundleProxy::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $this->companyUserFacadeMock = $this->getMockBuilder(CompanyUserFacadeInterface::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $this->dependencyProvider = new CustomerCompanyUserConnectorDependencyProvider();
-    }
-
-    /**
-     * @return void
-     */
-    public function testProvideBusinessLayerDependencies(): void
-    {
-        $this->containerMock->expects(static::atLeastOnce())
-            ->method('getLocator')
-            ->willReturn($this->locatorMock);
-
-        $this->locatorMock->expects(static::atLeastOnce())
-            ->method('__call')
-            ->withConsecutive(['companyUser'])
-            ->willReturn($this->bundleProxyMock);
-
-        $this->bundleProxyMock->expects(static::atLeastOnce())
-            ->method('__call')
-            ->withConsecutive(['facade'])
-            ->willReturnOnConsecutiveCalls(
-                $this->companyUserFacadeMock,
-            );
-
-        $container = $this->dependencyProvider->provideBusinessLayerDependencies($this->containerMock);
-
-        static::assertEquals($this->containerMock, $container);
-
-        static::assertInstanceOf(
-            CustomerCompanyUserConnectorToCompanyUserFacadeInterface::class,
-            $container[CustomerCompanyUserConnectorDependencyProvider::FACADE_COMPANY_USER],
-        );
-    }
-
-    /**
-     * @return void
-     */
-    public function testProvidePersistenceLayerDependencies(): void
-    {
-        $container = $this->dependencyProvider->providePersistenceLayerDependencies($this->containerMock);
-
-        static::assertEquals($this->containerMock, $container);
-
-        static::assertInstanceOf(
-            SpyCompanyUserQuery::class,
-            $container[CustomerCompanyUserConnectorDependencyProvider::QUERY_COMPANY_USER],
-        );
-    }
-}
