+++ codeception.yml
@@ -11,7 +11,7 @@ include:
   - bundles/conditional-availability-page-search-rest-api
   - bundles/conditional-availability-page-search-rest-api-extension
   - bundles/conditional-availability-sales-connector
-  - bundles/customer-company-user-connector
+  - bundles/customer-anonymizer-company-user-connector
   - bundles/jellyfish-sales-order-conditional-availability
   - bundles/product-list-conditional-availability-page-search
