+++ composer.json
@@ -27,7 +27,7 @@
     "fond-of-impala/conditional-availability-page-search-rest-api": "*",
     "fond-of-impala/conditional-availability-page-search-rest-api-extension": "*",
     "fond-of-impala/conditional-availability-sales-connector": "*",
-    "fond-of-impala/customer-company-user-connector": "*",
+    "fond-of-impala/customer-anonymizer-company-user-connector": "*",
     "fond-of-impala/jellyfish-sales-order-conditional-availability": "*",
     "fond-of-impala/product-list-conditional-availability-page-search": "*"
   },
@@ -86,7 +86,7 @@
       "FondOfImpala\\Shared\\ConditionalAvailabilityPageSearchRestApiExtension\\": "bundles/conditional-availability-page-search-rest-api-extension/src/FondOfImpala/Shared/ConditionalAvailabilityPageSearchRestApiExtension",
       "FondOfImpala\\Shared\\ConditionalAvailabilitySalesConnector\\": "bundles/conditional-availability-sales-connector/src/FondOfImpala/Shared/ConditionalAvailabilitySalesConnector",
       "FondOfImpala\\Zed\\ConditionalAvailabilitySalesConnector\\": "bundles/conditional-availability-sales-connector/src/FondOfImpala/Zed/ConditionalAvailabilitySalesConnector",
-      "FondOfImpala\\Zed\\CustomerCompanyUserConnector\\": "bundles/customer-company-user-connector/src/FondOfImpala/Zed/CustomerCompanyUserConnector",
+      "FondOfImpala\\Zed\\CustomerAnonymizerCompanyUserConnector\\": "bundles/customer-anonymizer-company-user-connector/src/FondOfImpala/Zed/CustomerAnonymizerCompanyUserConnector",
       "FondOfImpala\\Shared\\JellyfishSalesOrderConditionalAvailability\\": "bundles/jellyfish-sales-order-conditional-availability/src/FondOfImpala/Shared/JellyfishSalesOrderConditionalAvailability",
       "FondOfImpala\\Zed\\JellyfishSalesOrderConditionalAvailability\\": "bundles/jellyfish-sales-order-conditional-availability/src/FondOfImpala/Zed/JellyfishSalesOrderConditionalAvailability",
       "FondOfImpala\\Client\\ProductListConditionalAvailabilityPageSearch\\": "bundles/product-list-conditional-availability-page-search/src/FondOfImpala/Client/ProductListConditionalAvailabilityPageSearch",
@@ -120,7 +120,7 @@
       "FondOfImpala\\Shared\\ConditionalAvailabilityPageSearchRestApiExtension\\": "bundles/conditional-availability-page-search-rest-api-extension/tests/FondOfImpala/Shared/ConditionalAvailabilityPageSearchRestApiExtension",
       "FondOfImpala\\Shared\\ConditionalAvailabilitySalesConnector\\": "bundles/conditional-availability-sales-connector/tests/FondOfImpala/Shared/ConditionalAvailabilitySalesConnector",
       "FondOfImpala\\Zed\\ConditionalAvailabilitySalesConnector\\": "bundles/conditional-availability-sales-connector/tests/FondOfImpala/Zed/ConditionalAvailabilitySalesConnector",
-      "FondOfImpala\\Zed\\CustomerCompanyUserConnector\\": "bundles/customer-company-user-connector/tests/FondOfImpala/Zed/CustomerCompanyUserConnector",
+      "FondOfImpala\\Zed\\CustomerAnonymizerCompanyUserConnector\\": "bundles/customer-anonymizer-company-user-connector/tests/FondOfImpala/Zed/CustomerAnonymizerCompanyUserConnector",
       "FondOfImpala\\Shared\\JellyfishSalesOrderConditionalAvailability\\": "bundles/jellyfish-sales-order-conditional-availability/tests/FondOfImpala/Shared/JellyfishSalesOrderConditionalAvailability",
       "FondOfImpala\\Zed\\JellyfishSalesOrderConditionalAvailability\\": "bundles/jellyfish-sales-order-conditional-availability/tests/FondOfImpala/Zed/JellyfishSalesOrderConditionalAvailability",
       "FondOfImpala\\Client\\ProductListConditionalAvailabilityPageSearch\\": "bundles/product-list-conditional-availability-page-search/tests/FondOfImpala/Client/ProductListConditionalAvailabilityPageSearch",
