+++ dandelion.json
@@ -43,8 +43,8 @@
       "path": "bundles/conditional-availability-sales-connector",
       "version": "1.0.0"
     },
-    "customer-company-user-connector": {
-      "path": "bundles/customer-company-user-connector",
+    "customer-anonymizer-company-user-connector": {
+      "path": "bundles/customer-anonymizer-company-user-connector",
       "version": "1.0.0"
     },
     "jellyfish-sales-order-conditional-availability": {
